refactor data type in profit model.

+++ lib/features/profit/data/models/profit_model.dart
@@ -1,10 +1,10 @@
 class ProfitModel {
   int? acID;
   int? level;
-  int? mony;
+  double? mony;
   bool? creditORDepit;
   String? description;
-  int? acIndex;
+  double? acIndex;
   bool? isLast;
 
   ProfitModel(
